Throw an exception if tableIndexPathForSection: is invalid
Formerly, we used an NSAssert to check that the section was in the table. Since asserts are generally disabled for release builds, this meant that this method would return a bogus value (NSNotFound, or MAXINT) if the section wasn't found, leading to a weird crash.

This changes the API so that it will raise an NSInternalConsistencyException if the section is not a member of the table view, and will crash with a clearer error. If you plan to check for the possibility of a `nil` return, you can use the new form, tableIndexPathForSection:index:mustExist with mustExist set to NO.